+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 ------------------
 
 - added `safe_deepcopy` method to idc.api._utils which creates a deep copy of an object if not None
-- added `rgb-to-grayscale` filter
-- added `sort-pixels` filter
+- added `rgb-to-grayscale` filter to convert color images into gray scale ones
+- added `sort-pixels` filter for grayscale images
 
 
 0.0.6 (2025-02-26)